chore: repackage packages for scoped modules
This adds the Apache 2.0 License and symlinks it to every package,
which may have to be changed in the future to have a hard copy.
Preferably, the LICENSE will get copied into the packaged npm
package in the build process.

This namespaces all of the packages with @atjson, prepping this
for work in the open.

I also did some work to improve our build process by fiddling with
lerna. Some notes from this:

- `lerna bootstrap --hoist` will not work with ember-cli since
  broccoli doesn't use the node package resolution strategy.
  Because of this, packages installed at the root of the monorepo
  are being looked up in the wrong location.