🐛 Close any open annotations from Google Docs paste (#92)
When pasting from a GDocs document, the annotations don't get closed
until the character _after_ the style is done. This means that after
we loop through all of the styles, we need to close any that are
currently open to retain the intended paste.

Google Docs adds underlines automatically to all links. This removes it from the converter to offset