Make Renderer.render a static method 🚨: Inheriting classes should call render staticly
This updates the base hir renderer to have a static render method,
so inheriting renderers are updated to follow suit.

This also adds the `document` to the hir renderer `Context` interface,
so that we have access to the source document while rendering out
annotations, which may be necessary if that rendering depends on
text that is unrelated to the hir tree.