🐝: Fix bug for rendering invalid delimiter runs (#103)
* 🐝: Fix bug for rendering invalid delimiter runs
See https://spec.commonmark.org/0.28/#delimiter-run

Markdown requires that a left-flanking delimiter run (ie, one that opens
emphasis) must:
- Not be followed by whitespace
- Be preceded by whitespace or punctuation if followed by punctuation

Similarly a right-flanking delimiter run (ie, one that closes emphasis)
must:
- Not be preceded by whitespace
- Be followed by whitespace or punctuation if preceded by punctuation

We were already correctly pushing boundary whitespace out of emphasized
text. This fix addresses the second requirements by checking if the
text immediately inside the boundary is punctuation and if the text
immediately outside is not whitespace or punctuation, and pushing the
punctuation outside the boundary in that case.