🐝 Preserve leading spaces when escaping OL md text
When escaping text that looks like a markdown ordered list, we want to
preserve leading whitespace. This fixes a bug where a sentence ending
with an italic and then a number would have the space between them
removed, as the end of the sentence is rendered as text but looks like
ordered list markdown:
`Ending with *italic* 1. ` -> `Ending with *italic*1\.`
We still have to escape the terminal period, but now it will render as
`Ending with *italic* 1. ` -> `Ending with *italic* 1\.` with the space
preserved.