🐝: fix consecutive line breaks
This updates the commonmark renderer to use backslash + newline to
represent a line break. This allows for multiple consecutive line breaks
which the old md (space + space + newline) did not.

This also fixes behavior for line breaks ending paragraphs and block
elements, as well as line breaks in code and html blocks.