🐞 Fix bug in converting linebreaks split by 1 char
We were relying on a complicated regex which looked for:
  1. a line start or non-newline character
  2. a single newline
  3. a line end or non-newline character
This regex would work except when the matches would overlap, which is
when the non-newline character in 3. would be the non-newline character
in 1. of a subsequent match. In those cases, the next match would be
missed.

It is far simpler just to keep track of which newlines we identify as
part of paragraph boundaries and then just skip them when identifying
linebreaks.