🐞 Fix bug when a linebreak is at end of a bold
A recent change to Google Docs was made where a bold that ends a line
may have the newline character included in the bold. When this happens,
we create a Linebreak annotation nested within the Bold. This results
in the Bold ending with a backslash + newline. In this case, we were
moving the newline out of the Bold but not the backslash, since this is
a punctuation character which is allowed to be on the inner boundary of
a delimiter run if the outer boundary is a whitespace character, like a
newline. However, since the backslash is not escaped, it will function
to escape the delimiter, resulting in bad markdown. The fix is to
always move unescaped backslashes out.