chore: 👋 bye conventional-changelog-emoji
this has become a bit of a pain point for developers working in the repository and correctly identifying features / fixes / etc.the goal of the emojis was to be more accessible for folks who aren't primarily devs full-time, or come from a non-technical background. it didn't succeed at achieving those goals, and we can make the repository more accessible in places where it counts more: documentation and tools!