feat(video annotations): add provider attribute to video embeds (#457)
Since atjson is already going through the trouble of normalizing these URLs and figuring out the providers, we wanted to take it a step further on consuming applications and add a provider attribute that will allow us to conditionally render based on the video provider. This is mostly relevant for AMP, where we need to use a different video component per provider source.