feat: implement slice rendering in React
this is a small rewrite of the React renderer
to be better optimized for performance.

we don't need to use an iterator here, and
can directly loop over the HIR, which is
being done directly

keys for nodes should now also be guaranteed
to be unique, which was previously sometimes
not the case due to splits

type definitions have also been updated and
fixed for React 18 TypeScript clients, which
changed the behavior of FC to exclude
the children prop.