fix: handle four and five star formatting of markdown
for a _very_ specific case, we need to invert
the nesting of bold and italic, since markdown
can't handle four / five star delimiter runs
when parsing.

previously, we were producing invalid
markdown that looked like the following:

```
***Space*****: the final frontier**
```

the five star run is invalid, and using a naïve
approach, we could adjust this so we fix the
italic mark to render as an underscore, which
would produce:

```
**_Space_****: the final frontier**
```

but this is also an invalid delimiter run!

we have to re-orient the bold and italic marks
in the tree to create the following markdown,
which is the only valid structure:

```
*__Space__***: the final frontier**
```

another solution here would be healing the
mark ranges, which would produce less
markup, and the same visual result:

```
***Space*: the final frontier**
```

we are choosing not to handle this right now,
since these _are_ cases that may come up
when markdown is written