fix: sorting start/end tokens for zero-length marks
Fix a bug where we were sorting the MARK_END token before the
MARK_START token for zero-length marks. In the case we have:
* two tokens with the same index
* and the same annotation ID
we assume that one of them is the start and the other is the end
token. If the first is the start token, return -1 in the comparator
to sort it first