fix: remove zero-length annotations in gdocs converter
This is a bit of a workaround to manage edge cases that our
serializer does not handle well that mainly arise in GDocs and not
in a lot of other formats.

Specifically, there is a known issue where the tokens generated
for zero-length annotations can be sorted in non-deterministic
ways during document serialization. This can create marks with
malformed ranges if the end token is sorted ahead of the start
token for that annotation.

For now, we can remove these annotations to sidestep this problem