fix: malformed ranges in 0-length marks
Fix an issue where under certain circumstances we were producing
malformed ranges for 0-length marks. This would happen if the
MARK_END token was sorted before the MARK_START token for that
annotation, which could only happen for 0-length marks that
coincided with a PARSE_END and/or BLOCK_START. Even in that case,
the error depended on the way the list of tokens was divided and
merged during the internals of the sorting algorithm.

The fix here sidesteps the issue by replacing the MARK_START and
MARK_END tokens with a single MARK_COLLAPSED token in this case.
The mark range can be easily computed from the token since we know
that it is collapsed.