feat: rename alignment -> textAlignment and combine size and inset into the layout attribute
alignment was unclear as to what it was referring to and textAlignment
works better in different contexts.

layout is a new property that is in place of size and inset to be
more extensible for designers, and to be more explicit that mixing
size with block alignment (the inset property) doesn't actually work.
this allows us to create a more consistent property that is available
on more blocks without having this limited approach where size is on
one thing, inset is on another, and some have both. having a single
attribute for describing this will allow downstream consumers to handle
this more easily.