feat: attribute standardization (#1685)
The current attributes for our standard annotations have some messiness about them. This is an effort to clean up attributes and to provide a more standardized API that is consistent across more annotations. These decisions are made in conjunction with thinking about how our content is presented by our designers for sites like vogue.com, where some features that editors and writers use are not representable currently in our annotations.

We would like to provide more support and therefore, I'm introducing two major changes:

size and inset become layout.
These fields did not mix and match and allowing both properties made it confusing for developers as to which field took precedence. This also allows designers to add new types of layouts without having to change the data model.
alignment becomes textAlignment
This is pretty self-explanatory, and is about clarity.
Note that layout as a property is added to many more blocks than size and inset were on previously. This is intentional, since there are instances where editorial teams have used these features on these blocks. We may not provide them controls to change these properties, but we would like to fully support these and respect their decisions when it comes time to represent the data for our sites to use for rendering.

Additionally, support was added for x.com / twitter.com, and other unimplemented embeds were removed.