chore: change linebreak md syntax and dont remove them
Instead of suppressing the rendering of linebreaks when they
terminate a block/document, we can change our linebreak md syntax
to an alternative that doesn't break the parser.

The alternative linebreak sequence is `  \n` instead of `\\\n`.
When this appears at the end of a document, it is still parsed
as a linebreak, so it is fine to always render it.