fix: handle indents for blockquotes in gdocs-paste converter (#1880)
There is an issue with the converter where, if text has nested indentation, we end up creating nested blockquotes which are not supported in the rich text editor. We never want to do this, and instead want to ignore any additional indentation - ie, text that is further indented is just treated as normal text in the blockquote.